+++ app/src/main/java/io/github/yawnoc/strokeinput/StrokeInputService.java
@@ -78,4 +78,8 @@ public class StrokeInputService
     }
   }
   
+  @Override
+  public void onLongPress(String valueText) {
+  }
+  
 }